# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Campus Echo, a nearby anonymous campus feed.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /radar/presence:
    put:
      operationId: updateRadarPresence
      tags: [radar]
      summary: Opt in to nearby visibility or refresh the current presence
      description: Stores only the latest rounded location while opted in. Presence expires quickly and is deleted when hidden.
      security:
        - ClerkSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarPresenceInput"
      responses:
        "200":
          description: Presence is visible to nearby eligible users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarPresenceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/OutsideGeofence"
        "422":
          description: GPS accuracy is not sufficient for the 100-meter radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: hideRadarPresence
      tags: [radar]
      summary: Immediately remove the current user's radar presence
      security:
        - ClerkSessionCookie: []
      responses:
        "204":
          description: Presence removed
        "401":
          $ref: "#/components/responses/Unauthorized"
  /radar/nearby:
    get:
      operationId: getNearbyRadarBlips
      tags: [radar]
      summary: Find opted-in active users within the 100-meter radar
      description: Returns only opaque blip tokens, coarse direction sectors, and distance bands. Exact coordinates and account identities are never returned.
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: query
          name: latitude
          required: true
          schema:
            type: number
            minimum: -90
            maximum: 90
        - in: query
          name: longitude
          required: true
          schema:
            type: number
            minimum: -180
            maximum: 180
        - in: query
          name: accuracyMeters
          required: true
          schema:
            type: number
            minimum: 0
            maximum: 1000
      responses:
        "200":
          description: Coarse nearby presence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarNearbyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/OutsideGeofence"
        "422":
          description: GPS accuracy is not sufficient for the 100-meter radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /radar/inbox:
    get:
      operationId: getRadarInbox
      tags: [radar]
      summary: Get incoming and outgoing Ping requests and anonymous chats
      security:
        - ClerkSessionCookie: []
      responses:
        "200":
          description: Current radar inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarInbox"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /radar/socket-ticket:
    post:
      operationId: createRadarSocketTicket
      tags: [radar]
      summary: Create a one-use short-lived ticket for the authenticated WebSocket connection
      security:
        - ClerkSessionCookie: []
      responses:
        "201":
          description: Single-use WebSocket ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarSocketTicketResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /radar/chats/{chatId}/messages:
    get:
      operationId: listRadarChatMessages
      tags: [radar]
      summary: Load recent messages for an accepted anonymous chat
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Recent chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarMessagesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: sendRadarChatMessage
      tags: [radar]
      summary: Send a message to the other participant in an accepted anonymous chat
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarMessageInput"
      responses:
        "201":
          description: Message saved and sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          description: Message limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /radar/blips/{blipId}/block:
    post:
      operationId: blockRadarBlip
      tags: [radar]
      summary: Block an anonymous radar participant
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: blipId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Participant blocked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /radar/blips/{blipId}/report:
    post:
      operationId: reportRadarBlip
      tags: [radar]
      summary: Privately report an anonymous radar participant
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: blipId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarReportInput"
      responses:
        "201":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /radar/chats/{chatId}/block:
    post:
      operationId: blockRadarChatParticipant
      tags: [radar]
      summary: Block the other anonymous chat participant
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Participant blocked and chat closed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /radar/chats/{chatId}/report:
    post:
      operationId: reportRadarChatParticipant
      tags: [radar]
      summary: Privately report the other anonymous chat participant
      security:
        - ClerkSessionCookie: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarReportInput"
      responses:
        "201":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ClerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
  responses:
    Unauthorized:
      description: Sign in to continue
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    OutsideGeofence:
      description: The user is outside the 2 km radius of any campus hub
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GeofenceError"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested item does not exist or has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    GeofenceError:
      type: object
      properties:
        error:
          type: string
        distanceKm:
          type: number
        radiusKm:
          type: number
      required: [error, radiusKm]
    CampusHub:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radiusKm:
          type: number
      required: [id, name, city, country, latitude, longitude, radiusKm]
...
    RadarPresenceInput:
      type: object
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracyMeters:
          type: number
          minimum: 0
          maximum: 1000
      required: [latitude, longitude, accuracyMeters]
    RadarPresenceResponse:
      type: object
      properties:
        visible:
          type: boolean
        accuracyMeters:
          type: integer
        expiresInSeconds:
          type: integer
      required: [visible, accuracyMeters, expiresInSeconds]
    RadarNearbyResponse:
      type: object
      properties:
        hub:
          $ref: "#/components/schemas/CampusHub"
        radiusMeters:
          type: integer
        accuracyMeters:
          type: integer
        blips:
          type: array
          items:
            $ref: "#/components/schemas/RadarBlip"
        checkedAt:
          type: string
          format: date-time
      required: [hub, radiusMeters, accuracyMeters, blips, checkedAt]
    RadarBlip:
      type: object
      properties:
        blipId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [N, NE, E, SE, S, SW, W, NW]
        distanceBand:
          type: string
          enum: ["0-25m", "26-50m", "51-75m", "76-100m"]
      required: [blipId, direction, distanceBand]
    RadarSocketTicketResponse:
      type: object
      properties:
        ticket:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [ticket, expiresAt]
...
    RadarInbox:
      type: object
      properties:
        incomingPings:
          type: array
          items:
            $ref: "#/components/schemas/RadarPing"
        outgoingPings:
          type: array
          items:
            $ref: "#/components/schemas/RadarPing"
        chats:
          type: array
          items:
            $ref: "#/components/schemas/RadarChatSummary"
      required: [incomingPings, outgoingPings, chats]
    RadarPing:
      type: object
      properties:
        pingId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, accepted, declined, expired]
        direction:
          type: string
          enum: [incoming, outgoing]
        chatId:
          type: ["string", "null"]
          format: uuid
      required: [pingId, createdAt, expiresAt, status, direction, chatId]
    RadarChatSummary:
      type: object
      properties:
        chatId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastMessage:
          type: ["string", "null"]
      required: [chatId, createdAt, updatedAt, lastMessage]
    RadarMessagesResponse:
      type: object
      properties:
        chatId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: "#/components/schemas/RadarMessage"
      required: [chatId, messages]
    RadarMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
          maxLength: 500
        sentAt:
          type: string
          format: date-time
        fromMe:
          type: boolean
      required: [id, text, sentAt, fromMe]
    RadarMessageInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
      required: [text]
    RadarReportInput:
      type: object
      properties:
        reason:
          type: string
          enum: [harassment, unsafe, spam, other]
        details:
          type: string
          maxLength: 200
      required: [reason]
    RadarActionResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]